+++ src/components/alert/alert.ts
@@ -2,4 +2,4 @@ import SlAlert from './alert.component';
 export * from './alert.component';
 
 export default SlAlert;
-window.customElements.define('sl-alert', SlAlert);
+window.customElements.define('sl-alert', class extends SlAlert {});
